Switch to paris (#357)
* Changing log level of connected peers to DEBUG instead of info

* Switch to Paris based on TTD as a part of the delegated consensus

* Switching based on TTD for simplicity

* More fixes

* Fixed maru genesis file

* Fixed the ElDelegated consensus fork id serialization

+++ maru/consensus/src/main/kotlin/maru/consensus/ProtocolStarter.kt
@@ -99,7 +99,7 @@ class ProtocolStarter(
   }
 
   private fun checkAndHandleForkTransition() {
-    val currentTimestamp = clock.instant().epochSecond
+    val currentTimestamp = clock.instant().epochSecond.toULong()
     val nextBlockTimestamp = nextBlockTimestampProvider.nextTargetBlockUnixTimestamp(currentTimestamp)
     val nextForkSpec = forksSchedule.getForkByTimestamp(nextBlockTimestamp)
